Restructuring the ProtocolEmulator
Details:
	- Creating new files and structuring the code into these
	- Adding the testOffline() test
