+++ .gitignore
@@ -7,4 +7,4 @@ cppcheck/
 documentation/
 RDB_Diplomaterv_Monitor.pro.user*
 tests/test_config.json
-
+tests/protocol_emulator/emulators/__pycache__/*
